Fix test_execute

+++ test/test_execute.py
@@ -44,7 +44,8 @@ class TestExecute(unittest.TestCase):
 
     def tearDown(self):
         for f in self.temp_files:
-            file_target(f).unlink()
+            if file_target(f).exists():
+                file_target(f).unlink()
 
     def touch(self, files):
         '''create temporary files'''
@@ -416,7 +417,8 @@ input: group_by=2
         '''Test option paired_with '''
         self.touch(['a.txt', 'b.txt'])
         for ofile in ['a.txt1', 'b.txt2']:
-            file_target(ofile).unlink()
+            if file_target(ofile).exists():
+                file_target(ofile).unlink()
         #
         # string input
         script = SoS_Script(r'''
@@ -472,7 +474,8 @@ run: expand=True
         '''Test option group_with '''
         self.touch(['a.txt', 'b.txt'])
         for ofile in ['a.txt1', 'b.txt2']:
-            file_target(ofile).unlink()
+            if file_target(ofile).exists():
+                file_target(ofile).unlink()
         #
         # string input
         script = SoS_Script(r'''
@@ -937,8 +940,9 @@ with open('b.txt', 'w') as txt:
 
     def testRemovedIntermediateFiles(self):
         '''Test behavior of workflow with removed internediate files'''
-        file_target('a.txt').unlink()
-        file_target('aa.txt').unlink()
+        for file in ('a.txt', 'aa.txt'):
+            if file_target(file).exists():
+                file_target(file).unlink()
         script = SoS_Script('''
 [10]
 output: 'a.txt'
@@ -964,7 +968,7 @@ run: expand=True
         Base_Executor(wf).run()
         #
         # now we request the generation of target
-        file_target('a.txt').remove('target')
+        file_target('a.txt').unlink()
         file_target('aa.txt').unlink()
         Base_Executor(wf).run()
         #
@@ -974,7 +978,8 @@ run: expand=True
     def testStoppedOutput(self):
         '''test output with stopped step'''
         for file in ["{}.txt".format(a) for a in range(10)]:
-            file_target(file).unlink()
+            if file_target(file).exists():
+                file_target(file).unlink()
 
         script = SoS_Script('''
 [test_1]
@@ -1001,7 +1006,8 @@ assert(len(_input) == 5)
 
     def testAllowError(self):
         '''Test option allow error'''
-        file_target('a.txt').remove('all')
+        if file_target('a.txt').exists():
+            file_target('a.txt').unlink()
         script = SoS_Script('''
 [test]
 run:  allow_error=True
@@ -1013,7 +1019,7 @@ run:
         wf = script.workflow()
         Base_Executor(wf).run()
         self.assertTrue(file_target('a.txt').target_exists())
-        file_target('a.txt').remove('all')
+        file_target('a.txt').unlink()
 
     def testConcurrentWorker(self):
         '''Test the starting of multiple workers #493 '''
@@ -1031,7 +1037,8 @@ input: for_each={'i': range(2)}
     def testDependsCausedDependency(self):
         # test for #674
         for tfile in ('1.txt', '2.txt', '3.txt'):
-            file_target(tfile).unlink()
+            if file_target(tfile).exists():
+                file_target(tfile).unlink()
         script = SoS_Script('''
 [1: shared = {'dfile':'_output'}]
 output: '1.txt'
@@ -1054,7 +1061,8 @@ run: expand=True
         Base_Executor(wf).run()
         for tfile in ('1.txt', '2.txt', '3.txt'):
             self.assertTrue(file_target(tfile).target_exists())
-            file_target(tfile).unlink()
+            if file_target(tfile).exists():
+                file_target(tfile).unlink()
 
     def testConcurrentInputOption(self):
         '''Test input option'''
@@ -1169,8 +1177,9 @@ depends: sos_step('step')
 
     def testMultiSoSStep(self):
         '''Test matching 'a_1', 'a_2' etc with sos_step('a')'''
-        file_target('a_1').unlink()
-        file_target('a_2').unlink()
+        for file in ('a_1', 'a_2'):
+            if file_target(file).exists():
+                file_target(file).unlink()
         script = SoS_Script('''
 [a_b_1]
 output: "a_1"
@@ -1195,8 +1204,9 @@ depends: sos_step('a_b')
 
     def testDependsAuxiAndForward(self):
         '''Test depends on auxiliary, which then depends on a forward-workflow #983'''
-        file_target('a_1').unlink()
-        file_target('a_2').unlink()
+        for f in ('a_1', 'a_2'):
+            if file_target(f).exists():
+                file_target(f).unlink()
         script = SoS_Script('''
 
 [hg_1]
@@ -1224,8 +1234,9 @@ depends: "a_2"
 
     def testDependsAuxiAndSingleStepForward(self):
         '''Test depends on auxiliary, which then depends on a single-step forward-workflow'''
-        file_target('a_1').unlink()
-        file_target('a_2').unlink()
+        for f in ('a_1', 'a_2'):
+            if file_target(f).exists():
+                file_target(f).unlink()
         script = SoS_Script('''
 
 [hg_1]
@@ -1251,7 +1262,8 @@ depends: "a_2"
 
     def testDryrunPlaceholder(self):
         '''Test the creation and removal of placeholder files in dryrun mode'''
-        file_target('1.txt').unlink()
+        if file_target('1.txt').exists():
+            file_target('1.txt').unlink()
         script = SoS_Script('''
 a = '1.txt'
